requirements.txt
Django==5.0.2
django-redis==5.4.0
django-celery-results==2.5.1
djangorestframework==3.14.0
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0

